Adds verification check endpoint to change a user's `email verified` status (Server)
Adds an endpoint which takes an `email` string and an `authcode` string, which are compared against the database, rejecting if either are not equal.
Adds a function to update a given users `verified` flag field in the database, given some email.
Changes the build file to fail if the `go build` command fails, to catch errors before testing.