Fixes to `DentistAvailability` and `DentistInfo` components (Client)
Fixes `DentistAvailability` in the case of no availability being found.
Ensures DentistID is given as a number.
Changes `DentistInfo` component so that only filled address fields are displayed, in the case where only part of the address is known.
Adds `availability` to the `ResponseMessage` type in `authServices.ts` to fix a TypeScript error.